Fixes for type conversion warnings on visual studio
-Also possibly patched a bug where the code would crash with visual studio
in *some* cases (somewhere in the R source after my code returned).

+++ chordalGraphCommon/conditionalPoisson.cpp
@@ -145,7 +145,7 @@ beginSample:
 		//We start by computing the normalising constants. First index is k, second is z. All indices in this loop are 1 indexed
 		expNormalisingConstant.resize(nUnits,n);
 		//This will skip over the *ignored* units (the ones that were deterministically included)
-		int k = expExponentialParameters.size();
+		int k = (int)expExponentialParameters.size();
 		for(int unitIndex = nUnits; unitIndex >= 1; unitIndex--)
 		{
 			while(ignore[k-1]) k--;
